Remove use of `time` crate in favour of using `std::time`
See #14 for the lead-up to this change.

This adds a new `unix_time::Instant` type which fills the role
previously occupied by the `time::Timespec` type. Specifically:

- Retrieving current unix time.
- Converting between `ShortFormat` and `TimestampFormat` types.

However, note that the `std::fmt::Display` implementation has been
removed as a result. Instead, the documentation suggests the user
considers the `chrono` crate if they desire conversions to or from more
human-readable formats. A doc example is provided demonstrating how to
pretty print a `unix_time::Instant` using the `chrono` crate. This
allows us to remove the crate dependency and keep the ntp crate scope a
little more narrow.

Closes #14.