Remove unnecessary main.rs file from `src/`
It looks like this may have been left over from some previous work that
used ntp within an executable. I imagine that if we want to start work
on an ntp dameon in the future it might be better off in its own
subdirectory? Either way, best to remove this file until it's actually
used/needed.